move code from lib.__init__.py to new parsing.py
